+++ bundle/Form/Type/ExportType.php
@@ -6,6 +6,7 @@ use eZ\Publish\API\Repository\Repository;
 use Netgen\Bundle\InformationCollectionBundle\API\Value\Export\ExportCriteria;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\DataMapperInterface;
+use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\DateType;
 use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
@@ -36,6 +37,19 @@ class ExportType extends AbstractType
             'constraints' => [
                 new Assert\NotBlank(),
                 new Assert\Date(),
+            ],
+        ]);
+
+        $builder->add('exportType', ChoiceType::class, [
+            'required' => true,
+            'choices' => [
+                'netgen_information_collection_admin_export_type_csv' => 'csv',
+                'netgen_information_collection_admin_export_type_xls' => 'xls',
+            ],
+            'translation_domain' => 'netgen_information_collection_admin',
+            'label' => 'netgen_information_collection_admin_export_type',
+            'constraints' => [
+                new Assert\NotBlank(),
             ],
         ]);